backend/requirements.txt
Django
django-cors-headers
django-extensions
djangorestframework
matplotlib
PyJWT
six
psycopg2
python-decouple
